my_env/pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "my-env-sql-debugger"
version = "1.0.0"
description = "OpenEnv SQL debugger environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
]

[project.scripts]
server = "my_env.server.app:main"

[tool.setuptools]
packages = ["my_env", "my_env.server"]

[tool.setuptools.package-dir]
my_env = "."
"my_env.server" = "server"
